mega_env removed to its own file.
